Use data from the clock fuzzer
