+++ apycula/tiled_fuzzer.py
@@ -101,172 +101,45 @@ params = {
         "device": "GW1NS-2C-LQFP144-5",
         "partnumber": "GW1NS-UX2CLQ144C5/I4",
         "recode_idx": recode_idx_gw1ns_2,
-        "has_bottom_quadrants": False,
-        "lw_taps": [1, 0, 3, 2],
     },
     "GW1NS-4": {
         "package": "QFN48",
         "device": "GW1NSR-4C-QFN48-7",
         "partnumber": "GW1NSR-LV4CQN48PC7/I6",
         "recode_idx": recode_idx_gw1ns_4,
-        "has_bottom_quadrants": False,
-        "lw_taps": [2, 1, 0, 3],
     },
     "GW1N-9": {
         "package": "PBGA256",
         "device": "GW1N-9-PBGA256-6",
         "partnumber": "GW1N-LV9PG256C6/I5",
         "recode_idx": recode_idx_gw1n9,
-        "has_bottom_quadrants": True,
-        "lw_taps": [3, 2, 1, 0],
     },
     "GW1N-9C": {
         "package": "UBGA332",
         "device": "GW1N-9C-UBGA332-6",
         "partnumber": "GW1N-LV9UG332C6/I5",
         "recode_idx": recode_idx_gw1n9,
-        "has_bottom_quadrants": True,
-        "lw_taps": [3, 2, 1, 0],
     },
     "GW1N-4": {
         "package": "PBGA256",
         "device": "GW1N-4-PBGA256-6",
         "partnumber": "GW1N-LV4PG256C6/I5",
         "recode_idx": recode_idx_gw1n4,
-        "has_bottom_quadrants": False,
-        "lw_taps": [2, 1, 0, 3],
     },
     "GW1N-1": {
         "package": "LQFP144",
         "device": "GW1N-1-LQFP144-6",
         "partnumber": "GW1N-LV1LQ144C6/I5",
         "recode_idx": recode_idx_gw1n1,
-        "has_bottom_quadrants": False,
-        "lw_taps": [1, 0, 3, 2],
     },
     "GW1NZ-1": {
         "package": "QFN48",
         "device": "GW1NZ-1-QFN48-6",
         "partnumber": "GW1NZ-LV1QN48C6/I5",
         "recode_idx": recode_idx_gw1nz_1,
-        "has_bottom_quadrants": False,
-        "lw_taps": [1, 0, 3, 2],
     },
 }[device]
 
-def get_bufs_bits(fse, ttyp, win, wout):
-    wi = clknumbers[win]
-    wo = clknumbers[wout]
-    fuses = []
-    for rec in fse[ttyp]['wire'][38]:
-        if rec[0] == wi and rec[1] == wo:
-            fuses = chipdb.unpad(rec[2:])
-    return {fuse_h4x.fuse_lookup(fse, ttyp, f) for f in fuses}
-
-# create aliases and pipes for long wires
-def make_lw_aliases(fse, dat, db):
-    # type 81, 82, 83, 84 tiles have source muxes
-    center_row, col82 = dat['center']
-    center_row -= 1
-    last_row = db.rows - 1
-    col82 -= 1
-    col81 = col82 - 1
-    col83 = col82 + 1
-    col84 = col82 + 2
-    # type 91 and 92 tiles activate the quadrants
-    # XXX GW1NS-4 have different types
-    type91 = fse['header']['grid'][61][0][col82]
-    type92 = fse['header']['grid'][61][last_row][col82]
-    col91 = col82
-
-    # quadrants activation bels
-    # delete direct pips long wire -> spine because the artificial bel will be used,
-    # which replaces this direct pip
-    rows = {(0, 'T', type91)}
-    if params['has_bottom_quadrants']:
-        rows.update({ (last_row, 'B', type92) })
-    for row, half, ttyp in rows:
-        for idx in range(8):
-            bel = db.grid[row][col91].bels.setdefault(f'BUFS{idx}', chipdb.Bel())
-            del db.grid[row][col91].clock_pips[f'LWSPINE{half}L{idx}']
-            del db.grid[row][col91].clock_pips[f'LWSPINE{half}R{idx}']
-            src = f'LW{half}{idx}'
-            db.grid[row][col91].clock_pips.setdefault(f'LWI{idx}', {})[src] = {}
-            db.grid[row][col91].clock_pips.setdefault(f'LWSPINE{half}L{idx}', {})[f'LWO{idx}'] = {}
-            bel.flags['L'] = get_bufs_bits(fse, ttyp, f'LW{half}{idx}', f'LWSPINE{half}L{idx}')
-            db.grid[row][col91].clock_pips.setdefault(f'LWSPINE{half}R{idx}', {})[f'LWO{idx}'] = {}
-            bel.flags['R'] = get_bufs_bits(fse, ttyp, f'LW{half}{idx}', f'LWSPINE{half}R{idx}')
-            bel.portmap['I'] = f'LWI{idx}'
-            bel.portmap['O'] = f'LWO{idx}'
-            # aliases for long wire origins (center muxes)
-            # If we have only two quadrants, then do not create aliases in the bottom tile 92,
-            # thereby excluding these wires from the candidates for routing
-            if half == 'B' and not params['has_bottom_quadrants']:
-                continue
-            if half == 'T':
-                if idx != 7:
-                    db.aliases.update({(row, col82, src) : (center_row, col82, src)})
-                else:
-                    db.aliases.update({(row, col82, src) : (center_row, col81, src)})
-            else:
-                if idx != 7:
-                    db.aliases.update({(row, col82, src) : (center_row, col83, src)})
-                else:
-                    db.aliases.update({(row, col82, src) : (center_row, col84, src)})
-    # branches
-    # {lw#: {tap_col: {cols}}
-    taps = {}
-    lw_taps = params['lw_taps']
-
-    for lw in range(4):
-        tap_col = lw_taps[lw]
-        for col in range(db.cols):
-            if (col > tap_col + 2) and (tap_col + 4 < db.cols):
-                tap_col += 4
-            taps.setdefault(lw, {}).setdefault(tap_col, set()).add(col)
-
-    for row in range(db.rows):
-        for lw, tap_desc in taps.items():
-            for tap_col, cols in tap_desc.items():
-                tap_row = 0
-                if row > (center_row * 2):
-                    tap_row = last_row
-                db.aliases.update({(row, tap_col, 'LT01') : (tap_row, tap_col, 'LT02')})
-                db.aliases.update({(row, tap_col, 'LT04') : (tap_row, tap_col, 'LT13')})
-                for col in cols:
-                    db.aliases.update({(row, col, f'LB{lw}1') : (row, tap_col, f'LBO0')})
-                    db.aliases.update({(row, col, f'LB{lw + 4}1') : (row, tap_col, f'LBO1')})
-
-    # tap sources
-    rows = { (0, 'T') }
-    if params['has_bottom_quadrants']:
-        rows.update({ (last_row, 'B') })
-    for row, qd in rows:
-        for lw, tap_desc in taps.items():
-            for tap_col, cols in tap_desc.items():
-                if tap_col <= col91:
-                    half = 'L'
-                else:
-                    half = 'R'
-                db.aliases.update({ (row, tap_col, 'SS00') : (row, col91, f'LWSPINE{qd}{half}{lw}') })
-                db.aliases.update({ (row, tap_col, 'SS40') : (row, col91, f'LWSPINE{qd}{half}{lw + 4}') })
-                # XXX remove all pips except SS00 and SS40
-                pip2keep = {'SS00', 'SS40'}
-                for tp in ['LT02', 'LT13']:
-                    for pip in [p for p in db.grid[row][tap_col].pips[tp] if p not in pip2keep]:
-                        del db.grid[row][tap_col].pips[tp][pip]
-
-
-    # logic entries
-    srcs = {}
-    for i, src in enumerate(dat['UfbIns']):
-        row, col, pip = src
-        if pip == 126: # CLK2
-            db.aliases.update({ (center_row, col82, f'UNK{i + 104}'): (row - 1, col -1, 'CLK2')})
-            db.aliases.update({ (center_row, col81, f'UNK{i + 104}'): (row - 1, col -1, 'CLK2')})
-            if params['has_bottom_quadrants']:
-                db.aliases.update({ (center_row, col83, f'UNK{i + 104}'): (row - 1, col -1, 'CLK2')})
-                db.aliases.update({ (center_row, col84, f'UNK{i + 104}'): (row - 1, col -1, 'CLK2')})
 
 
 name_idx = 0
@@ -1258,9 +1131,6 @@ if __name__ == "__main__":
         bel.portmap['XXX_VSS1'] = wirenames[dat[f'IologicAIn'][42]]
     chipdb.dat_aliases(dat, db)
     chipdb.diff2flag(db)
-
-    # long wires
-    make_lw_aliases(fse, dat, db)
 
     # must be after diff2flags in order to make clean mask for OPEN_DRAIN
     fse_open_drain(fse, db, pin_locations)
